Upd: deb-setup.sh better handling of files
Add: config .tmux.conf

+++ extra/zsh.sh
@@ -1,7 +1,11 @@
 #!/bin/bash
 cd "$(dirname "$0")"
 
-sh -c "$(curl -fsSL https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh)" --unattended --keep-zshrc
+mkdir -p ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins}
 git clone https://github.com/zsh-users/zsh-autosuggestions ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-autosuggestions
 git clone https://github.com/zsh-users/zsh-syntax-highlighting.git ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-syntax-highlighting
-ln -sr ../.zshrc ~/
+if [[ ! -f ~/.zshrc ]]; then
+    ln -sr ../.zshrc ~/
+fi
+
+sh -c "$(curl -fsSL https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh)" --unattended --keep-zshrc
